refact: seperate header files and change header include in .c files, and rename .c files by feature
