+++ include/execute.h
@@ -6,7 +6,7 @@
 /*   By: kyusulee <kyusulee@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/30 17:36:52 by kyusulee          #+#    #+#             */
-/*   Updated: 2024/01/30 17:46:22 by kyusulee         ###   ########.fr       */
+/*   Updated: 2024/01/30 19:40:17 by kyusulee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 # include "../kyusulib/kyusulib.h"
 # include "minishell.h"
+# include "list.h"
 # include <fcntl.h>
 # include <sys/wait.h>
 
@@ -24,6 +25,7 @@
 # define OUTFILE_T_O (1)
 # define OUTFILE_A_O (2)
 # define DEFAULT_PATH "/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/bin:."
+
 
 typedef struct s_info
 {
@@ -47,5 +49,8 @@ void		children_switch(t_info *info, int *fd, t_cmd_node *cmd, char **env);
 // access.c
 void		exec(t_exe_lst *exes, char **env);
 char		*get_cmd(char *cmd, char **env, int flag);
+
+// execute_test.c
+void		execute_test(char **env);
 
 #endif
