+++ include/ui.h
@@ -1,20 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   traverse.h                                         :+:      :+:    :+:   */
+/*   ui.h                                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: younghoc <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: kyusulee <kyusulee@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/01/30 15:52:31 by younghoc          #+#    #+#             */
-/*   Updated: 2024/01/30 19:48:54 by kyusulee         ###   ########.fr       */
+/*   Created: 2024/01/30 19:20:34 by kyusulee          #+#    #+#             */
+/*   Updated: 2024/01/30 19:43:25 by kyusulee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef TRAVERSE_H
-# define TRAVERSE_H
+#ifndef UI_H
+# define UI_H
 
-# include <parse.h>
+# include <stdio.h>
 
-void	traverse(t_tree *tree, char **envp);
+// ui_frankshell_image.c
+void		print_frankshell_image(void);
 
 #endif
