+++ src/parse.c
@@ -6,7 +6,7 @@
 /*   By: younghoc <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/29 13:39:56 by younghoc          #+#    #+#             */
-/*   Updated: 2024/01/29 13:40:03 by younghoc         ###   ########.fr       */
+/*   Updated: 2024/01/30 19:54:25 by kyusulee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
